Initial mostly working add functions for list.h
Although the code is absolutely disgusting and needs cleaning
multiple times over, for now, this is a progress save.

There are two add functions which will likely be condensed to one
add function with more editing. As well, the code will eventually
be set up to allow for structs of different shapes and sizes
(with some restrictions to allow for how generic the list is).

The eventual goal is for these list methods to support linking structs
of any type that follow some basic structure.