add functions condensed; tested with struct ptr
With regards to the add functions that were separated by one method
that would only add to head and another that would only add after
having added to the head, both functions are now a single one,
with the extra comments removed.

As well, now that the add methods are more functional, I ran a very
basic test with struct pointers instead of structs.