Tested on string list
To test that the list add function works overall, I've begun testing
add with other structs like a string containing struct. As well
I'm beginning to replace some of the arguments to add with more
accurate values (the starter values happened to be correct by chance
as a result of padded space, but were not the right values to use).