list works for all orders of struct variables
Modified the listtest.c file so that rather than using the address
of some values in the struct to determine the offset, instead, use
the starting address of the struct to determine the offset of the
addresses of next and (optional) previous pointers.

Since the offset will be the same for all structs used in a linkedlist
of any type T, the offset can be used universally between all elements
of a given list.